requirements.txt
bidict==0.22.1
certifi @ file:///C:/ci_311/certifi_1676421629969/work/certifi
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
gTTS==2.3.1
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
pygame==2.3.0
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
urllib3==1.26.15
wincertstore==0.2
